Fix several Psalm issues

+++ lib/SimpleSAML/Utils/Random.php
@@ -24,6 +24,6 @@ class Random
      */
     public function generateID(): string
     {
-        return '_' . bin2hex(openssl_random_pseudo_bytes((int) ((self::ID_LENGTH - 1) / 2)));
+        return '_' . bin2hex(openssl_random_pseudo_bytes((self::ID_LENGTH - 1) / 2));
     }
 }
